refactor(api): package corrections
