+++ api/src/main/java/com/znaji/event_ticket_platform/api/events/io/AddTicketTypeRequest.java
@@ -1,11 +1,11 @@
-package com.znaji.event_ticket_platform.api.events;
+package com.znaji.event_ticket_platform.api.events.io;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
 import java.math.BigDecimal;
 
-public record UpdateTicketTypeRequest(
+public record AddTicketTypeRequest(
         @NotBlank String name,
         @NotNull BigDecimal price,
         @NotNull int maxQuantity
