+++ api/src/main/java/com/znaji/event_ticket_platform/api/events/io/TicketTypeResponse.java
@@ -1,4 +1,4 @@
-package com.znaji.event_ticket_platform.api.events;
+package com.znaji.event_ticket_platform.api.events.io;
 
 import java.math.BigDecimal;
 import java.util.UUID;
