+++ api/src/main/java/com/znaji/event_ticket_platform/api/events/io/UpdateEventRequest.java
@@ -1,13 +1,11 @@
-package com.znaji.event_ticket_platform.api.events;
+package com.znaji.event_ticket_platform.api.events.io;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
 import java.time.LocalDateTime;
-import java.util.UUID;
 
-public record CreateEventRequest(
-        @NotNull UUID organizerId,
+public record UpdateEventRequest(
         @NotBlank String name,
         String description,
         @NotNull LocalDateTime start,
