+++ api/src/main/java/com/znaji/event_ticket_platform/application/events/commands/AddTicketTypeCommand.java
@@ -1,11 +1,10 @@
-package com.znaji.event_ticket_platform.application.events;
+package com.znaji.event_ticket_platform.application.events.commands;
 
 import java.math.BigDecimal;
 import java.util.UUID;
 
-public record UpdateTicketTypeCommand(
+public record AddTicketTypeCommand(
         UUID eventId,
-        UUID ticketTypeId,
         String name,
         BigDecimal price,
         int maxQuantity
