+++ api/src/main/java/com/znaji/event_ticket_platform/application/events/commands/EventIdCommand.java
@@ -1,4 +1,4 @@
-package com.znaji.event_ticket_platform.application.events;
+package com.znaji.event_ticket_platform.application.events.commands;
 
 import java.util.UUID;
