+++ api/src/main/java/com/znaji/event_ticket_platform/application/events/commands/UpdateTicketTypeCommand.java
@@ -1,10 +1,11 @@
-package com.znaji.event_ticket_platform.application.events;
+package com.znaji.event_ticket_platform.application.events.commands;
 
 import java.math.BigDecimal;
 import java.util.UUID;
 
-public record AddTicketTypeCommand(
+public record UpdateTicketTypeCommand(
         UUID eventId,
+        UUID ticketTypeId,
         String name,
         BigDecimal price,
         int maxQuantity
